Desktop/khareetaty-ai-mvp/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
pandas
scikit-learn
plotly
